+++ packages/context/src/AddressIdentity.tsx
@@ -1,6 +1,6 @@
 import { AccountDisplay } from "@polkadot-api/react-components";
 import { FC } from "react";
-import { useIdentity } from "./polkahubContext";
+import { useIdentity, useSS58Formatter } from "./polkahubContext";
 
 export const AddressIdentity: FC<{
   addr: string;
@@ -9,12 +9,13 @@ export const AddressIdentity: FC<{
   className?: string;
   maxAddrLength?: number;
 }> = ({ addr, name, className, maxAddrLength = 12, copyable = true }) => {
+  const formatSS58 = useSS58Formatter();
   let identity = useIdentity(addr);
 
   return (
     <AccountDisplay
       account={{
-        address: addr,
+        address: formatSS58(addr),
         name: identity?.name ?? name,
         subId: identity?.subId,
         verified: identity?.verified,