+++ packages/plugin/src/index.ts
@@ -1,2 +1,3 @@
 export type * from "./plugin";
 export * from "./persist";
+export * from "./addr";